requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.14
PyYAML==5.1
validators==0.13.0